1st pass Guard Builder
