+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tempo-monorepo",
-  "version": "2.9.1",
+  "version": "2.9.2",
   "private": true,
   "description": "Magma Computing Monorepo",
   "repository": {
